+++ src/Rings/mpoly-graded.jl
@@ -2507,3 +2507,9 @@ end
 Return the ideal in the underlying ungraded ring.
 """
 forget_grading(I::MPolyIdeal{<:MPolyDecRingElem}) = forget_decoration(I)
+
+gens(A::MPolyDecRing, i::Int) = A[i]
+
+### This is a temporary fix that needs to be addressed in AbstractAlgebra, issue #1105.
+# TODO: This still seems to be not resolved!!!
+Generic.ordering(S::MPolyDecRing) = :degrevlex
